refactor(folder): 'classes' -> 'scripts'
