+++ scripts/network_segmentation.py
